fix: satisfy PET test address analyzer

+++ tests/PetEmulator.Pet.Tests/PetMemoryBusTests.cs
@@ -43,10 +43,10 @@ public sealed class PetMemoryBusTests
         bus.Read(0x3000).Should().Be(0xFF);
     }
 
-    [TestCase(0xE800)]
-    [TestCase(0xE830)]
-    [TestCase(0xE850)]
-    [TestCase(0xE890)]
+    [TestCase((ushort)0xE800)]
+    [TestCase((ushort)0xE830)]
+    [TestCase((ushort)0xE850)]
+    [TestCase((ushort)0xE890)]
     public void UnmappedPeripheralAddress_ReadsAsOpenBus(ushort address)
     {
         var bus = CreateBus(PetProfileCatalog.Pet2001_8, out _, out _, out _, out _);
